Refactor: Triangulation split into partial classes and changed namespace to ShapeEngine.Geometry.TriangulationDef.
