+++ Examples/Scenes/ExampleScenes/EndlessSpaceExampleSource/AsteroidObstacle.cs
@@ -7,6 +7,7 @@ using ShapeEngine.Geometry.CollisionSystem;
 using ShapeEngine.Geometry.PolygonDef;
 using ShapeEngine.Geometry.RectDef;
 using ShapeEngine.Geometry.TriangleDef;
+using ShapeEngine.Geometry.TriangulationDef;
 using ShapeEngine.StaticLib;
 using ShapeEngine.Random;
