+++ Examples/Scenes/ExampleScenes/PathfinderExample2.cs
@@ -11,6 +11,7 @@ using ShapeEngine.Geometry.CollisionSystem;
 using ShapeEngine.Geometry.PolygonDef;
 using ShapeEngine.Geometry.RectDef;
 using ShapeEngine.Geometry.TriangleDef;
+using ShapeEngine.Geometry.TriangulationDef;
 using ShapeEngine.Input;
 using ShapeEngine.Pathfinding;
 using Color = System.Drawing.Color;
