+++ ShapeEngine/Geometry/FractureInfo.cs
@@ -1,3 +1,4 @@
+using ShapeEngine.Geometry.TriangulationDef;
 using ShapeEngine.StaticLib;
 
 namespace ShapeEngine.Geometry;
