+++ ShapeEngine/Geometry/PolygonDef/Polygon.cs
@@ -9,6 +9,7 @@ using ShapeEngine.Geometry.RectDef;
 using ShapeEngine.Geometry.SegmentDef;
 using ShapeEngine.Geometry.SegmentsDef;
 using ShapeEngine.Geometry.TriangleDef;
+using ShapeEngine.Geometry.TriangulationDef;
 using ShapeEngine.Random;
 using ShapeEngine.StaticLib;
 using Size = ShapeEngine.Core.Structs.Size;
